pyproject.toml
[tool.poetry]
name = "yolo-x-server"
version = "0.0.2"
description = ""
authors = ["kohei <koheitech001@gmail.com>"]
license = "Apach-2.0"
readme = "README.md"
packages = [{include = "yolo_x_server"}]

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.6.2"
attrs = "22.2.0"
certifi = "2022.12.7"
click = "8.1.3"
coloredlogs = "15.0.1"
exceptiongroup = "1.1.0"
flatbuffers = "23.1.21"
h11 = "0.14.0"
httpcore = "0.16.3"
humanfriendly = "10.0"
idna = "3.4"
iniconfig = "2.0.0"
mpmath = "1.2.1"
numpy = "1.24.2"
onnxruntime = "1.14.0"
opencv-python = "4.7.0.68"
packaging = "23.0"
pillow = "9.4.0"
pluggy = "1.0.0"
protobuf = "4.22.0"
pydantic = "1.10.5"
pyparsing = "3.0.9"
pytest = "7.2.1"
rfc3986 = "1.5.0"
six = "1.16.0"
sniffio = "1.3.0"
sympy = "1.11.1"
tomli = "2.0.1"
typing-extensions = "4.5.0"
coco-formatter = {git = "https://github.com/kevin-tofu/coco_formatter.git", rev = "v0.0.1"}
mediarouter = {git = "https://github.com/kevin-tofu/MediaRouter.git", rev = "v0.0.2"}
filerouter = {git = "https://github.com/kevin-tofu/FileRouter.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
